Added and implemented reorder function.
Doing testing on example implementation and was getting misordered
solutions. Realized that up to this point the end user was required
to reorder the solution vector using the order vector returned from
the lu_decomp function to get the correct order for the solution
vector.

Wrote the simple function to reorder accordingly and included as
part of the solve function so the solutions are automatically
returned in the correct order.